+++ src/components/create-account-flow/components/stepCounter.tsx
